+++ adapters/Facebook/FacebookAdapter/OpenBidding/GADFBRewardedInterstitialRenderer.h
@@ -1,4 +1,4 @@
-// Copyright 2019 Google LLC.
+// Copyright 2020 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -15,12 +15,8 @@
 #import <Foundation/Foundation.h>
 #import <GoogleMobileAds/GoogleMobileAds.h>
 
-@interface GADFBRewardedRenderer : NSObject
+#import "GADFBRewardedRenderer.h"
 
-/// Asks the receiver to load the ad configuration.
-- (void)loadRewardedAdForAdConfiguration:
-            (nonnull GADMediationRewardedAdConfiguration *)adConfiguration
-                       completionHandler:
-                           (nonnull GADMediationRewardedLoadCompletionHandler)completionHandler;
+@interface GADFBRewardedInterstitialRenderer : GADFBRewardedRenderer
 
 @end